Refactoring ... / Retirando bibliotecas desnecessarias no package.json
