+++ api/enum/content-type.js
@@ -1,0 +1,3 @@
+module.exports = {
+    APPLICATION_IMAGE: "application/image"
+}
