+++ api/service/codigo-barras-service.js
