+++ package.json
@@ -7,27 +7,14 @@
     "debug": "node index.js"
   },
   "dependencies": {
-    "body-parser": "~1.8.1",
+    "@sfd-br/util": "0.0.10",
+    "body-parser": "^1.19.0",
     "boleto-brasileiro-validator": "^1.0.3",
-    "convert-source-map": "^1.1.0",
-    "cookie-parser": "~1.3.3",
-    "cors": "^2.8.5",
-    "debug": "^2.0.0",
-    "express": "^4.9.8",
-    "express-helpers": "^1.3.2",
-    "express-load": "^1.1.14",
-    "express-session": "^1.12.1",
-    "method-override": "^2.3.5",
-    "mongo-cursor-pagination": "^7.0.0",
-    "mongodb": "latest",
-    "mongodb-core": "latest",
+    "express": "^4.17.1",
+    "method-override": "^3.0.0",
+    "mongoose": "^5.5.13",
     "monk": "latest",
-    "promise": "^7.1.1",
-    "properties-reader": "0.0.15",
-    "qrcode": "^1.3.0",
-    "request": "^2.85.0",
-    "request-promise": "^4.2.2",
-    "string": "^3.3.3"
+    "qrcode": "^1.3.0"
   },
   "devDependencies": {}
 }
